+++ TreasureHuntModel/TreasureHuntModel/src/main/java/treasurehunt/model/StepFactory.java
@@ -2,6 +2,6 @@ package treasurehunt.model;
 
 public interface StepFactory {
 	
-	public Step createInstance(String id, float latitude, float longitude);
+	public Step createInstance(String id, double latitude, double longitude);
 	
 }
